+++ doc/whats_new.rst
@@ -12,6 +12,10 @@ v1.1.0
 
 * Add ``ARIMA.plot_diagnostics`` method
 
+* Add new arg to ``ARIMA`` constructor: ``with_intercept`` (default is True).
+
+* New default for ``trend`` is no longer ``'c'``, it is ``None``.
+
 
 v1.0.0
 ------